Patch 2 (#7)
* [dtlaunch, b2] Change swipe direction to UpDown

Change swipe directions to move between pages from LeftRight to UpDown, as to match the page indicators. Implemented like 'natural scrolling'.

Also added function showClock() which is called on 'left-to-right' swipe.  Here I imagine the clock to be to the left of the launcher, if you do that then this also follows the natural scrolling principle.

* Update ChangeLog

* Update ChangeLog

* Update ChangeLog

* Update ChangeLog

* Update metadata.json