spotrem:try to debug timeout inside swipe handler
The timeout will run the first time I trigger the handler, and sometimes
a few time more. But very soon the code inside the timeout will stop
being run when the swipe handler is triggered. The other code inside the
swipe handler that's outside of the timeout will run fine.

I've tried to reproduce this bug in a simpler environment, but there it
behaved as I would have expected.